Refactoring test packages
