+++ alchemist/alchemist-incarnation-scafi/src/test/scala/it/unibo/alchemist/scafi/test/ScafiPrograms.scala
@@ -1,12 +1,12 @@
 /*******************************************************************************
  * Copyright (C) 2010-2018, Danilo Pianini and contributors listed in the main
  * project's alchemist/build.gradle file.
- * 
+ *
  * This file is part of Alchemist, and is distributed under the terms of the
  * GNU General Public License, with a linking exception, as described in the file
  * LICENSE in the Alchemist distribution's top directory.
  ******************************************************************************/
-package test.scafiprograms
+package it.unibo.alchemist.scafi.test
 
 import it.unibo.alchemist.implementation.nodes.NodeManager
 import it.unibo.alchemist.model.interfaces.Node
