+++ alchemist/alchemist-incarnation-scafi/src/test/scala/it/unibo/alchemist/scafi/test/TestScafiIncarnation.scala
@@ -1,3 +1,5 @@
+package it.unibo.alchemist.scafi.test
+
 /*******************************************************************************
  * Copyright (C) 2010-2018, Danilo Pianini and contributors listed in the main
  * project's alchemist/build.gradle file.
@@ -6,15 +8,14 @@
  * GNU General Public License, with a linking exception, as described in the file
  * LICENSE in the Alchemist distribution's top directory.
  ******************************************************************************/
-import it.unibo.alchemist.model.implementations.actions.RunScafiProgram
 import it.unibo.alchemist.model.ScafiIncarnation
+import it.unibo.alchemist.model.implementations.actions.RunScafiProgram
 import it.unibo.alchemist.model.implementations.environments.Continuous2DEnvironment
-import it.unibo.alchemist.model.implementations.reactions.{ChemicalReaction, Event}
+import it.unibo.alchemist.model.implementations.positions.Euclidean2DPosition
+import it.unibo.alchemist.model.implementations.reactions.Event
 import it.unibo.alchemist.model.interfaces.Reaction
 import org.apache.commons.math3.random.MersenneTwister
 import org.scalatest.{FunSuite, Matchers}
-import it.unibo.alchemist.model.interfaces.Position
-import it.unibo.alchemist.model.implementations.positions.Euclidean2DPosition
 
 class TestScafiIncarnation extends FunSuite with Matchers {
   private val INC = new ScafiIncarnation[Euclidean2DPosition]
